backend/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.3
django-auth-ldap==4.1.0
django-cors-headers==3.11.0
django-environ==0.9.0
django-filebrowser-no-grappelli==4.0.1
django-grappelli==3.0.3
django-phonenumber-field==6.1.0
django-tinymce==3.4.0
phonenumbers==8.12.48
Pillow==9.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-ldap==3.4.2
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
psycopg2-binary==2.9.
gunicorn==20.1.0